+++ src/cli.ts
@@ -30,30 +30,19 @@ const run = async (options: Options): Promise<Result<void, Failure[]>> => {
       const {GitHubActionsCache} = await import(
         './caching/github-actions-cache.js'
       );
-      const cacheResult = GitHubActionsCache.create(logger);
-      if (!cacheResult.ok) {
-        if (cacheResult.error.reason === 'invalid-usage') {
-          return {
-            ok: false,
-            error: [
-              {
-                script: options.script,
-                type: 'failure',
-                reason: 'invalid-usage',
-                message: cacheResult.error.message,
-              },
-            ],
-          };
-        } else {
-          const never: never = cacheResult.error.reason;
-          throw new Error(
-            `Internal error: unexpected cache result error reason: ${String(
-              never,
-            )}`,
-          );
-        }
+      const cacheResult = await GitHubActionsCache.create(logger);
+      if (cacheResult.ok) {
+        cache = cacheResult.value;
+      } else {
+        cache = undefined;
+        console.warn(
+          '⚠️ Error initializing GitHub cache. Caching is disabled for this run',
+        );
+        logger.log({
+          script: options.script,
+          ...cacheResult.error,
+        });
       }
-      cache = cacheResult.value;
       break;
     }
     case 'none': {